DROOLS-1701 move range() factory to CompiledFEELSemanticMappings
