+++ kie-dmn/kie-dmn-feel/src/main/java/org/kie/dmn/feel/runtime/impl/RangeImpl.java
@@ -16,14 +16,7 @@
 
 package org.kie.dmn.feel.runtime.impl;
 
-import java.time.Period;
-
-import org.kie.dmn.api.feel.runtime.events.FEELEvent;
-import org.kie.dmn.feel.lang.EvaluationContext;
-import org.kie.dmn.feel.lang.ast.RangeNode;
 import org.kie.dmn.feel.runtime.Range;
-import org.kie.dmn.feel.runtime.events.ASTEventBase;
-import org.kie.dmn.feel.util.Msg;
 
 public class RangeImpl
         implements Range {
@@ -33,38 +26,6 @@ public class RangeImpl
     private Comparable    lowEndPoint;
     private Comparable    highEndPoint;
 
-    public static RangeImpl of(EvaluationContext ctx, RangeBoundary lowBoundary, Object lowEndPoint, Object highEndPoint, RangeBoundary highBoundary) {
-        Comparable left = asComparable(lowEndPoint);
-        Comparable right = asComparable(highEndPoint);
-        if (left == null || right == null || !compatible(left, right)) {
-            ctx.notifyEvt( () -> new ASTEventBase(FEELEvent.Severity.ERROR, Msg.createMessage(Msg.INCOMPATIBLE_TYPE_FOR_RANGE, left.getClass().getSimpleName() ), null));
-            return null;
-        }
-        return new RangeImpl(
-                lowBoundary,
-                left,
-                right,
-                highBoundary);
-    }
-
-    private static boolean compatible(Comparable left, Comparable right) {
-        Class<?> leftClass = left.getClass();
-        Class<?> rightClass = right.getClass();
-        return leftClass.isAssignableFrom(rightClass)
-                || rightClass.isAssignableFrom(leftClass);
-    }
-
-    private static Comparable asComparable(Object s) {
-        if (s instanceof Comparable) {
-            return (Comparable) s;
-        } else if (s instanceof Period) {
-            // period has special semantics
-            return new RangeNode.ComparablePeriod((Period) s);
-        } else {
-            // FIXME report error
-            return null;
-        }
-    }
 
     public RangeImpl() {
     }
